--- next-app/Api/Receptions/Models/PostReceptionsModel.d.ts
@@ -1,21 +1,0 @@
-export interface PostReceptionsReqData {
-  register_date: string[],
-}
-
-export interface PostReceptionsResData {
-  data: [
-    {
-      reception_id: number;
-      user_name: string;
-      start: string;
-      end: string;
-      reserved: boolean;
-      canceled: boolean;
-    }
-  ]
-}
-export interface PostReceptionsApiResData {
-  ok: boolean;
-  res?: PostReceptionsResData;
-  errorText?: string;
-}
